efficientad/requirements.txt
# conda create -n effad python=3.8
# conda activate effad
# (Windows) pip install torch==1.13.0+cu116 torchvision==0.14.0+cu116 torchaudio==0.13.0+cu116 -f https://download.pytorch.org/whl/cu116/torch_stable.html
# (Ubuntu) pip install torch==1.13.0+cu117 torchvision==0.14.0+cu117 torchaudio==0.13.0 --extra-index-url https://download.pytorch.org/whl/cu117

# torch==1.13.0
# torchvision==0.14.0
tifffile==2021.7.30
tqdm==4.56.0
scikit-learn==1.2.2
numpy #==1.18.5
Pillow==7.0.0
scipy #==1.7.1
tabulate==0.8.7
tifffile==2021.7.30
tqdm==4.56.0
pandas
opencv-python